feat(account): refactor file-based account persistence
