--- Core/Sources/Account/ProcessInfo.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-extension ProcessInfo {
-    var isTestEnvironment: Bool { environment["XCTestSessionIdentifier"] != nil }
-}
-
-var isTestEnvironment: Bool { ProcessInfo.processInfo.isTestEnvironment }
