--- Core/Sources/Account/URL.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-extension URL {
-    static func documents(_ path: String = "") -> Self {
-        (isTestEnvironment ? Self.temporaryDirectory : .documentsDirectory).appending(path: path)
-    }
-}
