--- Core/Tests/AccountTests/ProcessInfoTests.swift
@@ -1,9 +1,0 @@
-@testable import Account
-import Testing
-import Foundation
-
-struct ProcessInfoTests {
-    @Test func isTestEnvironment() {
-        #expect(ProcessInfo.processInfo.isTestEnvironment == true)
-    }
-}
