--- Core/Tests/AccountTests/URLTests.swift
@@ -1,11 +1,0 @@
-@testable import Account
-import Testing
-import Foundation
-
-struct URLTests {
-    @Test func documents() {
-        #expect(URL.documents("Test.json").absoluteString.hasSuffix("/Test.json"))
-        #expect(URL.documents("Test.txt") == .temporaryDirectory.appending(path: "Test.txt"))
-        #expect(URL.documents() == .temporaryDirectory)
-    }
-}
